refactor: share common modules
