+++ src/shared/cacheManager.js
